fix: fix keyboard error
but this is stopgap, it will need fix more

Issue Tracker: #32
