+++ reddit-clone/Podfile
@@ -9,4 +9,5 @@ target 'reddit-clone' do
   # Pods for reddit-clone
   pod 'SideMenu', '~> 6.0'
   pod 'Tabman', '~> 2.11'
+  pod 'IQKeyboardManagerSwift'
 end
